Update FetchServiceResponseOperation to use either dataTask or downloadTask
